move comp/apm/etwtracer to comp/tracer/etwtracer
