+++ .github/CODEOWNERS
@@ -219,7 +219,6 @@
 /comp @DataDog/agent-shared-components
 /comp/aggregator @DataDog/agent-shared-components
 /comp/api @DataDog/agent-shared-components
-/comp/apm @DataDog/windows-agent
 /comp/checks @DataDog/agent-shared-components
 /comp/collector @DataDog/agent-metrics-logs
 /comp/core @DataDog/agent-shared-components
@@ -246,6 +245,7 @@
 /comp/forwarder/orchestrator @DataDog/agent-metrics-logs
 /comp/forwarder/orchestrator/orchestratorinterface @DataDog/agent-metrics-logs
 /comp/metadata/packagesigning @DataDog/agent-platform
+/comp/trace/etwtracer @DataDog/windows-agent
 /comp/etw @DataDog/windows-agent
 # END COMPONENTS
