+++ cmd/agent/subcommands/run/command_windows.go
@@ -13,10 +13,10 @@ import (
 	_ "expvar"         // Blank import used because this isn't directly used in this file
 	_ "net/http/pprof" // Blank import used because this isn't directly used in this file
 
-	apmetwtracer "github.com/DataDog/datadog-agent/comp/apm/etwtracer"
-	apmetwtracerimpl "github.com/DataDog/datadog-agent/comp/apm/etwtracer/impl"
 	"github.com/DataDog/datadog-agent/comp/collector/collector"
 	etwimpl "github.com/DataDog/datadog-agent/comp/etw/impl"
+	apmetwtracer "github.com/DataDog/datadog-agent/comp/tracer/etwtracer"
+	apmetwtracerimpl "github.com/DataDog/datadog-agent/comp/tracer/etwtracer/impl"
 
 	"github.com/DataDog/datadog-agent/comp/checks/winregistry"
 	winregistryimpl "github.com/DataDog/datadog-agent/comp/checks/winregistry/impl"
