--- comp/apm/bundle.go
@@ -1,9 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-//go:build windows
-
-package apmetwtracer
-
-// team: windows-agent
