+++ comp/trace/etwtracer/component.go
